feat: 添加外观与语言设置页及入口
Verification: Manual inspection shows: 1) SettingAppearanceActivity added and registered in Manifest; 2) SettingActivity now includes appearance import and click handler; 3) Layout for appearance added and integrated; 4) XML well-formed; 5) No extraneous whitespace; 6) Imports cleaned up.

Task-Guard: beta-appearance-review-20260905
