+++ app/src/leanback/AndroidManifest.xml
@@ -261,6 +261,10 @@
            android:screenOrientation="sensorLandscape" />
 
         <activity
+            android:name=".ui.activity.SettingAppearanceActivity"
+            android:configChanges="screenSize|smallestScreenSize|screenLayout|orientation"
+            android:screenOrientation="sensorLandscape" />
+        <activity
             android:name=".ui.activity.SettingAdActivity"
             android:configChanges="screenSize|smallestScreenSize|screenLayout|orientation"
             android:screenOrientation="sensorLandscape" />
